backend/requirements.txt
flask
flask-cors
mediapipe
opencv-python
pillow
numpy
